# techforce-pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: FieldCommand — Premium fire protection field service management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/schedule:
    get:
      operationId: getEmployeeSchedule
      tags: [employees]
      summary: Get employee jobs for a date range
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  # ── CUSTOMERS ────────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/jobs:
    get:
      operationId: getCustomerJobs
      tags: [customers]
      summary: All jobs for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /customers/{id}/invoices:
    get:
      operationId: getCustomerInvoices
      tags: [customers]
      summary: All invoices for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"

  # ── JOBS ─────────────────────────────────────────────────────────────────────
...
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, role, certType, salary, billableRate, shopDaysAllowed, shopDaysUsedYtd, utilizationPct, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
          enum: [suppression, sprinkler, extinguisher, helper, admin]
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        shopDaysUsedYtd:
          type: integer
        allowedTrainingDays:
          type: integer
        trainingDaysUsedYtd:
          type: integer
        utilizationPct:
          type: number
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
        notes:
          type: string
    EmployeeInput:
      type: object
      required: [name, role, certType, salary, billableRate]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
          enum: [suppression, sprinkler, extinguisher, helper, admin]
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        hireDate:
          type: string
        notes:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        shopDaysUsedYtd:
          type: integer
        allowedTrainingDays:
          type: integer
        trainingDaysUsedYtd:
          type: integer
        utilizationPct:
          type: number
        status:
          type: string
        notes:
          type: string
    Customer:
      type: object
      required: [id, name, type, contact, phone, email, address, contractStatus, revenueYtd, jobCount]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [Restaurant, School, Commercial, "Group Home", Condo, Government, Healthcare, Industrial]
        contact:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        contractStatus:
          type: string
          enum: [active, expiring, pending]
        contractExpiry:
          type: ["string", "null"]
        revenueYtd:
          type: number
        jobCount:
          type: integer
        extinguisherRate:
          type: ["string", "null"]
        suppressionRate:
          type: ["string", "null"]
        sprinklerRate:
          type: ["string", "null"]
        exitLightRate:
          type: ["string", "null"]
        notes:
          type: string
    CustomerInput:
      type: object
      required: [name, type, contact, phone, email, address]
      properties:
        name:
          type: string
        type:
          type: string
        contact:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        contractStatus:
          type: string
        contractExpiry:
          type: ["string", "null"]
        extinguisherRate:
          type: ["string", "null"]
        suppressionRate:
          type: ["string", "null"]
        sprinklerRate:
          type: ["string", "null"]
        exitLightRate:
          type: ["string", "null"]
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        contact:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        contractStatus:
          type: string
        contractExpiry:
          type: ["string", "null"]
        extinguisherRate:
          type: ["string", "null"]
        suppressionRate:
          type: ["string", "null"]
        sprinklerRate:
          type: ["string", "null"]
        exitLightRate:
          type: ["string", "null"]
        notes:
          type: string
    Job:
      type: object
      required: [id, customerId, customerName, employeeId, techName, serviceType, scheduledDate, scheduledTime, status, priority, revenue, certRequired]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        employeeId:
          type: ["integer", "null"]
        techName:
          type: string
        serviceType:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, return, reschedule, emergency, cancelled, non_compliant, will_return]
        priority:
          type: string
          enum: [low, medium, high, critical]
        revenue:
          type: number
        certRequired:
          type: string
        recurring:
          type: boolean
        recurringInterval:
          type: ["string", "null"]
        notes:
          type: string
        completedAt:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        nonComplianceReason:
          type: ["string", "null"]
        nonComplianceNotifiedAt:
          type: ["string", "null"]
...
    LineItem:
      type: object
      required: [id, service, description, qty, unitPrice, total]
      properties:
        id:
          type: integer
        service:
          type: string
        description:
          type: string
        qty:
          type: number
        unitPrice:
          type: number
        total:
          type: number
...
    Invoice:
      type: object
      required: [id, customerId, customerName, status, issued, due, lineItems, subtotal, total, notes]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        techId:
          type: ["integer", "null"]
        techName:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issued:
          type: string
        due:
          type: string
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxPct:
          type: number
        total:
          type: number
        notes:
          type: string
        reminderSent:
          type: boolean
...
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        techId:
          type: ["integer", "null"]
        due:
          type: string
        notes:
          type: string
        reminderSent:
          type: boolean
...